+++ src/Data/Morpheus/Schema/Directive.hs
@@ -1,6 +1,7 @@
-{-# LANGUAGE DeriveGeneric #-}
-{-# LANGUAGE TypeFamilies  #-}
-{-# LANGUAGE TypeOperators #-}
+{-# LANGUAGE DeriveGeneric     #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TypeFamilies      #-}
+{-# LANGUAGE TypeOperators     #-}
 
 module Data.Morpheus.Schema.Directive
   ( Directive(..)
@@ -9,10 +10,14 @@ module Data.Morpheus.Schema.Directive
 import           Data.Morpheus.Kind                                (KIND, OBJECT)
 import           Data.Morpheus.Schema.DirectiveLocation            (DirectiveLocation)
 import           Data.Morpheus.Schema.Internal.RenderIntrospection (InputValue)
+import           Data.Morpheus.Types.GQLType                       (GQLType (__typeName))
 import           Data.Text                                         (Text)
 import           GHC.Generics                                      (Generic)
 
 type instance KIND Directive = OBJECT
+
+instance GQLType Directive where
+  __typeName = const "__Directive"
 
 data Directive = Directive
   { name        :: Text