+++ src/Data/Morpheus/Schema/InputValue.hs
@@ -1,16 +1,23 @@
-{-# LANGUAGE DeriveGeneric #-}
-{-# LANGUAGE TypeFamilies  #-}
+{-# LANGUAGE DeriveGeneric     #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TypeFamilies      #-}
 
 module Data.Morpheus.Schema.InputValue
   ( InputValue(..)
   , createInputValueWith
   ) where
 
-import           Data.Morpheus.Kind (KIND, OBJECT)
-import           Data.Text          (Text)
+import           Data.Morpheus.Kind          (KIND, OBJECT)
+import           Data.Morpheus.Types.GQLType (GQLType (__typeName))
+import           Data.Text                   (Text)
+import           Data.Typeable               (Typeable)
 import           GHC.Generics
 
 type instance KIND (InputValue a) = OBJECT
+
+instance Typeable a => GQLType (InputValue a) where
+  __typeName = const "__InputValue"
 
 data InputValue t = InputValue
   { name         :: Text